Add community domain counts
I would like to get an overview of and keep track of what
the largest broadcast domains are. So to be able to later
sort by the (community,domain) tuple in Grafana.

(Though some exceptions for this are probably Aachen and Braunschweig,
if I recall correctly, who are using their own broadcast domain
division tricks, which are not visible in the nodes.json/meshviewer.json?)

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>